# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NOYA WEB API Documentation",
        "description": "API documentation for NOYA WEB application",
        "contact": {
            "email": "admin@noyaweb.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "NOYA WEB API Server"
        }
    ],
    "paths": {
...
        "/api/v1/stocks": {
            "get": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Get list of stocks",
                "description": "Returns list of stocks with pagination",
                "operationId": "getStocksList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100
                        }
                    },
                    {
                        "name": "merchant_id",
                        "in": "query",
                        "description": "Filter by merchant",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "article_id",
                        "in": "query",
                        "description": "Filter by article",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stocks retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "links": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Create new stock",
                "description": "Create a new stock entry",
                "operationId": "storeStock",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "merchant_id",
                                    "article_id"
                                ],
                                "properties": {
                                    "merchant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "article_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "last_action_type": {
                                        "type": "string",
                                        "enum": [
                                            "MANUALLY_ADD",
                                            "MANUALLY_WITHDRAW",
                                            "AUTO_ADD",
                                            "AUTO_WITHDRAW"
                                        ],
                                        "example": "MANUALLY_ADD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stock created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/stocks/{stock}/add": {
            "post": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Add stock quantity",
                "description": "Add quantity to existing stock with history tracking",
                "operationId": "addStock",
                "parameters": [
                    {
                        "name": "stock",
                        "in": "path",
                        "description": "Stock id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "action_type": {
                                        "type": "string",
                                        "enum": [
                                            "MANUALLY_ADD",
                                            "AUTO_ADD"
                                        ],
                                        "example": "MANUALLY_ADD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock added successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/stocks/{stock}/withdraw": {
            "post": {
                "tags": [
                    "Stocks"
                ],
                "summary": "Withdraw stock quantity",
                "description": "Withdraw quantity from existing stock with history tracking",
                "operationId": "withdrawStock",
                "parameters": [
                    {
                        "name": "stock",
                        "in": "path",
                        "description": "Stock id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity"
                                ],
                                "properties": {
                                    "quantity": {
                                        "type": "integer",
                                        "example": 25
                                    },
                                    "action_type": {
                                        "type": "string",
                                        "enum": [
                                            "MANUALLY_WITHDRAW",
                                            "AUTO_WITHDRAW"
                                        ],
                                        "example": "MANUALLY_WITHDRAW"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock withdrawn successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Stock withdrawn successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}